move UI feature out of commons.eclipse.pde plugin
creates a new commons.eclipse.pde.ui plugin for features requiring UI
