+++ plugins/org.eclipse.gemoc.commons.eclipse.pde.ui/META-INF/MANIFEST.MF
@@ -1,0 +1,24 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: GEMOC Commons Eclipse PDE UI
+Bundle-SymbolicName: org.eclipse.gemoc.commons.eclipse.pde.ui;singleton:=true
+Bundle-Version: 3.3.0.qualifier
+Require-Bundle: org.eclipse.core.runtime,
+ org.eclipse.gemoc.commons.eclipse;visibility:=reexport,
+ org.eclipse.jdt.core,
+ org.eclipse.jdt.ui,
+ org.eclipse.pde.ui;visibility:=reexport,
+ org.eclipse.ui.workbench,
+ org.eclipse.swt,
+ org.eclipse.jface,
+ org.eclipse.ui.ide,
+ org.eclipse.core.resources
+Bundle-RequiredExecutionEnvironment: JavaSE-11
+Bundle-ActivationPolicy: lazy
+Export-Package: org.eclipse.gemoc.commons.eclipse.pde.ui,
+ org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde,
+ org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde.ui,
+ org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde.ui.templates
+Bundle-Vendor: Eclipse GEMOC Project
+Automatic-Module-Name: org.eclipse.gemoc.commons.eclipse.pde
+Bundle-Activator: org.eclipse.gemoc.commons.eclipse.pde.ui.Activator
