+++ plugins/org.eclipse.gemoc.commons.eclipse.pde.ui/build.properties
@@ -11,4 +11,6 @@
 source.. = src/
 output.. = bin/
 bin.includes = META-INF/,\
-               .
+               .,\
+               schema/,\
+               plugin.xml
