+++ plugins/org.eclipse.gemoc.commons.eclipse.pde.ui/plugin.xml
