+++ plugins/org.eclipse.gemoc.commons.eclipse.pde.ui/src/org/eclipse/gemoc/commons/eclipse/pde/ui/Activator.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     Inria - initial API and implementation
  *******************************************************************************/
-package org.eclipse.gemoc.commons.eclipse.pde;
+package org.eclipse.gemoc.commons.eclipse.pde.ui;
 
 import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.osgi.framework.BundleContext;
@@ -19,7 +19,7 @@ import org.osgi.framework.BundleContext;
 public class Activator extends AbstractUIPlugin {
 
 	// The plug-in ID
-	public static final String PLUGIN_ID = "org.eclipse.gemoc.commons.eclipse.pde"; //$NON-NLS-1$
+	public static final String PLUGIN_ID = "org.eclipse.gemoc.commons.eclipse.pde.ui"; //$NON-NLS-1$
 
 	// The shared instance
 	private static Activator plugin;
