+++ plugins/org.eclipse.gemoc.commons.eclipse.pde.ui/src/org/eclipse/gemoc/commons/eclipse/pde/wizards/pages/pde/ui/ProjectTemplateApplicationOperation.java
