+++ plugins/org.eclipse.gemoc.commons.eclipse.pde/META-INF/MANIFEST.MF
@@ -6,21 +6,12 @@ Bundle-Version: 3.3.0.qualifier
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.gemoc.commons.eclipse;visibility:=reexport,
  org.eclipse.jdt.core,
- org.eclipse.jdt.ui,
- org.eclipse.pde.ui;visibility:=reexport,
- org.eclipse.ui.workbench,
- org.eclipse.swt,
- org.eclipse.jface,
- org.eclipse.ui.ide,
  org.eclipse.core.resources
-Bundle-RequiredExecutionEnvironment: JavaSE-1.7
+Bundle-RequiredExecutionEnvironment: JavaSE-11
 Bundle-ActivationPolicy: lazy
 Export-Package: org.eclipse.gemoc.commons.eclipse.pde,
  org.eclipse.gemoc.commons.eclipse.pde.classpath,
  org.eclipse.gemoc.commons.eclipse.pde.java,
- org.eclipse.gemoc.commons.eclipse.pde.manifest,
- org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde,
- org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde.ui,
- org.eclipse.gemoc.commons.eclipse.pde.wizards.pages.pde.ui.templates
+ org.eclipse.gemoc.commons.eclipse.pde.manifest
 Bundle-Vendor: Eclipse GEMOC Project
 Automatic-Module-Name: org.eclipse.gemoc.commons.eclipse.pde
